Power profile scripts
Both of these are set to run from the /root directory and they need to
be run as root.

in crontab for root I have

*/4 * * * * * /root/pwr_prof

So it samples every 4 minutes and only takes a snapshot if the battery
is charging or discharging.

I've been running the crontab script across the holiday with no problems.