dlo#7932: Fix failsafe script in the presence of pretty-boot.
This is more complicated than I'd hoped.  In the presence of pretty-boot
we're started with input coming in on tty2 but output happening on tty1.
The resolution is to have one script that checks to see whether we're
in pretty-boot and switches to VT2 if we are, and to move the script
that actually does input/output to "diskspacerecover".  On ugly-boot
it'll run on tty1 as normal, on pretty-boot it'll run on tty2.

A long-term fix is to have /etc/init.d/z-boot-anim-stop return us to
a more sane environment than this; it should set the controlling
terminal back to tty1 and chvt 1, but I wasn't able to make this work.