Fallback to en_US if the manufacturing locale is not supported by the OS
Suppose you are trying to run an OS build developed for Paraguay that
omits support for hi_IN on a laptop manufactured for India. Without this
patch, the locale would be set to hi_IN. Since hi_IN is missing, Sugar
falls back to C. This patch makes the locale fall back to en_US if the
laptop's manufacturing locale is not supported.