rework udev rule and keymap override scheme
the previous commit(s) attempted to replace a file owned by udev,
which of course rpm will forbid.  this commit changes the name of the
olpc map file, and addes a rule to our set of "late" rules which will
cause that file to override the map applied by the stock udev rule.