Add olpc-panelpwr-log
This script can be used to measure the amount of power delivered to
the battery from the external power input.  It puts the laptop into
the lowest power mode and then wakes up every $DELAY (normally 300)
seconds to take a reading from the battery.
It normally used for measuring the amount power sourced from a solar
panel.