xorg: add support for displaylink devices, tidy up inhibit suspend
We add support for displaylink devices, which are available in the
market, work well and are have an actively supported linux driver.

The udlfb driver registers the next available fb device, we use a
udev rule to add an easy-to-find symlink. The device can read EDID
and autoconfigure X to a reasonable mode, so we let that happen.

Along the way we generalize the code a bit, and tidy up the
inhibit suspend mechanism, making it more self-contained in
olpc-configure.

udlfb devices work with both fbdev and displaylink xorg drivers.
The displaylink driver is more efficient, using the damage extension
to update the display selectively. This is clearly not working well
(BZ#816671) so we fall back on fbdev which works pretty well.