olpc-configure: don't mark no-resize after resizing
Setting the no-resize flag after resizing is not ideal. Instead
it would be better to check if there is space to resize the partition,
and if not, don't do anything. This is exactly what resize2fs already
does, so we can just run resize2fs on every "normal" boot without
worry.

This is sensible/simple and will also help against any strange
"my disk wasn't resized" case such as the one in #10040.