olpc-configure: allow customisation of default keyboard/language (#6689)
There have been cases found in the field where overriding the
manufacturing data (used to set initial language/keyboard settings) is
desirable. Once in Nepal in #6689, and now in Canada, where almost
every school in the project has a different language/keyboard
preference.

Allow initial settings to be set in /etc/olpc-configure which override
the manufacturing data, when found.