Enable maliit on-screen keyboard where available (#12082)
As we no longer ship alternative IM modules, there is no need to hardcode
the simple module - that will now always be used. See #6945 for the
original reasoning for why this was needed.

Enable Maliit when the GTK3 module can be found, when no other special
considerations (e.g. xim, scim) apply.