olpc-configure: root configuration cleanup
Our root version marker will be wiped out on all upgrades, so
there is no need to actually write the version there. Just use
the presence of the file to determine whether root upgrades
are needed or not.

To half-support yum upgrades, the RPM spec file will now delete
this marker after installation.

Also remove the /etc/X11/xorg.conf cleanup - on upgrades we
always start from a clean root so there is no need for this
to be here.